+++ backend/src/lib/bounded-contexts/marketing/marketing/application/event-handlers/domain/index.ts
@@ -1,0 +1,3 @@
+import { TodoCompletionsIncrementedHandler } from './todo-completions-incremented.handler';
+
+export const StreamingDomainEventHandlers = [TodoCompletionsIncrementedHandler];
