+++ backend/src/lib/bounded-contexts/marketing/marketing/application/event-handlers/integration/index.ts
@@ -1,0 +1,9 @@
+import { TodoCompletedIntegrationEventHandler } from './todo-completed.integration-handler';
+import { UserEmailChangedIntegrationEventHandler } from './user-email-changed.integration-handler';
+import { UserRegisteredIntegrationEventHandler } from './user-registered.integration-handler';
+
+export const StreamingIntegrationEventHandlers = [
+  UserRegisteredIntegrationEventHandler,
+  TodoCompletedIntegrationEventHandler,
+  UserEmailChangedIntegrationEventHandler,
+];
