+++ backend/src/lib/bounded-contexts/marketing/marketing/constants.ts
@@ -2,6 +2,8 @@ export const EmailServicePortToken = Symbol('EmailServicePort');
 
 export const StreamingCommandBusToken = Symbol('StreamingCommandBusToken');
 
+export const StreamingDomainEventBusToken = Symbol('StreamingDomainEventBus');
+
 export const UserWriteRepoPortToken = Symbol('UserWriteRepoPort');
 
 export const NotificationTemplateReadRepoPortToken = Symbol(
