+++ backend/src/lib/bounded-contexts/todo/todo/application/event-handlers/integration/index.ts
@@ -1,0 +1,1 @@
+export const StreamingIntegrationEventHandlers = [];
